third_party/blink/web_tests/platform/mac-mac10.13/external/wpt/html/syntax/speculative-parsing/generated/document-write/svg-script-src.tentative.sub-expected.txt

This is a testharness.js-based test.
FAIL Speculative parsing, document.write(): svg-script-src Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nHost: web-platform.test:8001\nConnection: keep-alive\nUser-Agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/999.77.34.5 Safari/537.36\nAccept: */*\nReferer: http://web-platform.test:8001/html/syntax/speculative-parsing/generated/document-write/svg-script-src.tentative.sub.html\nAccept-Encoding: gzip, deflate\nAccept-Language: en-us,en\n\n"
Harness: the test ran to completion.
